requirements.txt
haikunator==2.1.0
python-dotenv==1.0.0
Requests==2.31.0
watchdog==3.0.0
